Arm: Speed up -1..1 soft clipping with Neon
If the signal exceeds -1..1 then, as error handling, the soft_clip
function forces the signal back into -1..1. This is problematic since
the search loop to find the next sample exceeding -1..1 is slow.
If cheap on the current platform, while doing -2..2 hardclipping we
can also detect if the signal never exceeds -1..1, avoiding the need
for a second search loop.

Change-Id: I7d751afc2d335765798ed7b48993c0d51ff843bd